Change the `PathGlobs`, `SnapshotSubset`, and `UrlToFetch` intrinsics to return `Digest` (#10449)
The engine knows how to go from `Digest -> Snapshot`, but not the other way around. By changing these three intrinsics, rules have much more flexibility. 

Now, the claim in https://www.pantsbuild.org/v2.0/docs/rules-api-file-system#core-abstractions-digest-and-snapshot that "The core building block is a Digest" is true.

--

This also renames `SnapshotSubet` to `DigestSubset`, given that it takes as input `Digest` and returns a `Digest`.