Rename `UrlToFetch` to `DownloadFile` (#10468)
Our `fs.py` types generally take the form of a declarative verb to indicate that it's a request, e.g. `AddPrefix` or `CreateDigest`. See https://www.pantsbuild.org/v2.0/docs/rules-api-file-system. This aligns with that.

This also emphasizes that you are downloading a single file/asset, rather than making a generic GET request.