Eliminate most rule registration boilerplate. (#10477)
Introduce a collect_rules function that can find all @rules and the
SubsystemRules they depend on in a given namespace. This can be used to
both reduce boilerplate and also make rule development less surprising.
Using collect_rules it now becomes rarer to edit rules and encounter
errors due to forgetting to register rules for the various introduced
subsystems and @rule functions.

Also streamline the Rule interface and the `def rules()` semantics;
now all rules are Rules save for UnionRule which can be dealt away
with later to leave users only exposed to registering Rules.