Make independent owners requests per file to improve memoization (#10491)
### Problem

#10441 caused a performance regression from about 10s to run `./pants dependencies --transitive ::` with inference enabled, to about 22s.

### Solution

Make independent owners requests per file we'd like to find owners, which allows the lookups for each file to be memoized independently.

### Result

`./pants dependencies --transitive ::` takes 15s. Although we are doing more work than before (before #10441, conftest discovery would only happen at `test` time), this is not ideal, and we should do further optimization before launch. But there are a few variables that will impact this soon that make it not the best time to optimize: 1) an intrinisic `PathGlobs->Paths` operation, 2) possibly enabling inference by default, allowing all of the strategies to use the module_mapper.

[ci skip-rust-tests]