Add cache_failures option to Process type (#10433)
### Problem

Failed processes are being memoized in pantsd (https://github.com/pantsbuild/pants/issues/10129 )

### Solution

This commit introduces a new flag on the `Process` type `cache_failures`, defaulting to False, which is in turn passed to to the engine's node-execution infrastructure. If this flag is false, the engine will treat a `MultiPlatformExecuteRequest` node as noncacheable rather than cacheable, which will cause such nodes to be re-run in the pantsd case mentioned in the above issue. Note that this involves changing the signature of the `Node::cacheable` method to now take into account the result value of a node.