Fix ExternalTool to use relative paths (#10503)
Before, we weren't prefixing the `exe` path with `./`, so the process would end up searching both for a local file and also on the `$PATH`. (This searching of `$PATH` is the same mechanism that allows us to run `python` for Python processes).

While this might not end up being an issue in most cases, it is less tight than we'd like. For example, it is possible for the user to have `pex` discoverable on their PATH already.

[ci skip-build-wheels]
[ci skip-rust]