Add `--pex-executable-search-path` (#10489)
Previously, `--python-setup-interpreter-search-paths` would be used both to find Python interpreters and to set the $PATH for the running PEX subprocess. This resulted in many awkward/impossible set ups.

Switch to using PEX_PYTHON_PATH to both discover a bootstrap
interpreter and steer runtime interpreter selection. Introduce a
find_binary rule to facilitate this that uses ~portable bash for now.
Also introduce a PexRuntimeEnvironment to steer bootstrap interpreter
probing and allow specification of the PATH that should be exposed to
the PEX runtime environment.

Fixes #9760.

[ci skip-rust]
[ci skip-build-wheels]