Restore the missing init injection logic.
However now we only inject if the missing files exist and
are empty. If they exist but are not empty, we error.

The restored logic is slightly changed from the one
we had previously - it no longer does injection, but
just finds the missing files. It's up to the caller
to merge digests if they so choose.

For consistency, renamed inject_ancestor_files.py to
ancestor_files.py, since that code just finds ancestor
files, but doesn't do any injection.

[ci skip-rust]

[ci skip-build-wheels]