Remove `tarutil.py` (#10516)
`tarutil_test.py` is failing with Toolchain CI and on macOS for Tom. We're only using this code now for the deprecated `binary_util.py`, which we'd like to finish replacing with `ExternalTool`.

Background: `tarutil.py` was added in https://github.com/pantsbuild/pants/commit/a9957e91d8f262c8d1384c380cf8fb77ec075eb1 due to Python 2's `tarutil.py` not reporting as many errors eagerly as we'd like. Now that we're using Python 3 with a more modern implementation—and we don't use `binary_util.py`—this doesn't seem worth the cost of keeping this code.

[ci skip-rust]
[ci skip-build-wheels]