Remove rust source include from integration tests. (#10522)
### Problem

Integration tests depend transitively on the `native_engine.so`, which is continuously validated because it must be present to run, and which is bootstrapped before pants starts up. But integration tests were also including the rust sources unnecessarily.

### Solution

Remove the dependency on the rust sources, and the associated target. Also, align the comment about our use of `pants_local_binary` to what we learned in #10279.

[ci skip-rust]
[ci skip-build-wheels]