docker image for remote execution without GCP/RBE (#10515)
### Problem

The ability of the Pants project to use RBE on GCP ends soon. The image used for remote execution relies on a GCP/RBE image.

### Solution

Create a similar image that does not rely on RBE/GCP.

### Result

Initial testing of this image internally with Toolchain's remote execution cluster shows that many tests pass. Further iteration may be necessary if some additional dependencies are missing that would have been included in the RBE/GCP image.