Support and require disambiguated file addresses (#10535)
This adds a new syntax for generated subtargets / file deps to specify which original base target they come from: `a/b/c.txt:original`, with the following characteristics:

* If the original target has a "default name", meaning the `name` field is left off, then there is no need for the disambiguation.
* As before, a base target can live in an ancestor directory, e.g. `a/b/c.txt:../../original`.

Thanks to this new syntax, we no longer look for the Owners of explicit file dependencies. We require the values to be unambiguous, which improves performance and removes an edge case where we would error if there were multiple owners of the file.

We use the engine now to parse an `AddressInput` into an `Address`. This is necessary so that we can properly determine if a path is a file or a directory, as `a/b/c` could technically be a file or a directory.

In followups, we can extend this syntax to file specs. It also will allow us to consistently use generated subtargets when using file specs, which will simplify lots of that code and resolve https://github.com/pantsbuild/pants/issues/10455.