Avoid Python 3.8+ for Pylint test (#10548)
### Problem

As described in https://github.com/pantsbuild/pants/issues/10547, success of `PylintIntegrationTest.test_uses_correct_python_version` depends on the Python version. The astroid package (v1.6.6) used by pylint is looking for an `ast` node that exists Python 3.6 but was removed in Python 3.8.

### Solution

Use an interpreter constraint to avoid Python 3.8+ for this test.

### Result

Test failure is gone.