remoting: use non-RBE pants remote execution image (#10539)
### Problem

GCP is ending usage of RBE for Pants on August 7th. The current remote execution image uses an RBE image as the base. Besides RBE going away, we will not be able to rely on this base image being available once the RBE alpha ends.

### Solution

Use a new image for remote execution which is similar to the RBE base image but is based on Ubuntu 16 Xenial directly plus any necessary dependencies for Pants that were assumed to be in the RBE base image.

### Result

All tests should pass.

[ci skip-rust]