Stop using `AddressWithOrigin` for precise file arguments (#10551)
We used to achieve precise file arguments—where you only run on the file you specified—via `AddressWithOrigin` et al., including the `SpecifiedSourceFiles` type.

Now that we always use a generated subtarget when a user gives us file args, we know that the target will always have exactly one source file, so we can simply run on the entire target's `sources` without needing to consider taking a subset of its sources.

This allows us to remove `SpecifiedSourceFiles`, along with storing which files were actually resolved on the `FilesystemSpec`. We also simplify `test`, `lint`, and `fmt` to stop using `WithOrigin`, as none of the downstream rules need that information anymore.

We still keep around `WithOrigin` for the sake of errors, such as allowing us to error if a user specifies a literal file arg and it does not meet an invariant, but to no-op if they specified a looser glob.