Ensure that cmd_line_spec_parser.py does the right thing
This leans in on using `SingleAddress`, rather than using a file arg. It would be awkward to get a file arg to work because those will try to find the `Snapshot` of the file, which is unnecessary. Also, conceptually the user is giving us an address, just a file address

# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]