Fix a regression in dependency inference + the dependees goal
Now that we store `--tag` and `--exclude-target-regexp` on `AddressMapper`, rather than `AddressSpecs`, we end up always using those values and always filtering. This is an issue in some places where we use `Get(Addresses, AddressSpecs)` to learn about the universe of targets - we don't want to filter in those cases.

We solve this by adding a field to `AddressSpecs` called `apply_target_filters`. It does not have a default value because we want the author to have to explicitly decide what behavior they want.

# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]