Add `--build-patterns` option (#10559)
This option allows you to use a different pattern for BUILD file names other than `['BUILD', 'BUILD.*']`. It appears that we always wanted it, but hadn't wired it up.

This also makes several refactors to the `AddressSpecs` code in preparation for adding support for disambiguated file addresses on the command line, like `foo.txt:tgt`.

* Removes `AddressSpecsMatcher` and inlines it into `AddressMapper`. There was no need for a separate type.
* Moves all tests for `base/` from `tests/python` to `src/python`.
* Refactor setting up the `PathGlobs` for `AddressSpecs`.
* Moves `more_specific()` to `AddressSpecs.more_specific()` and adds `FilesystemSpecs.more_specific()`.

This change would naively result in us always applying the filters `--tag` and `--exclude-target-regexp` anytime we us the rule to go from `AddressSpecs -> Addresses`, even though we sometimes don't want those filters. We solve this by adding a field `apply_target_filters: bool` to `AddressSpecs`.