A python_distribution target.
This is now the only target type that can have a provides= clause.

This simplifies the model for running setup-py, as all
python_library targets that are owned by a dist are now on equal
footing (whereas previously one of them would appear to be
privileged, as the one with the provides=clause).

[ci skip-rust]

[ci skip-build-wheels]