Allow for disabling pantsrc in OptionsBootstrapper independently of options values (#10562)
### Problem

`pantsrc` file consumption is enabled by default via options, but many usages of `OptionsBootstrapper` forget to disable it in tests. Because `pantsrc` resolution uses absolute file lookups, it is able to break out of the test sandbox, which can make for some confusing errors. 

### Solution

Disable `pantsrc` consumption by default at the `OptionsBootstrapper` level, and then enable it in production usecases, or usecases that intentionally change the default `pantsrc` locations in order to test that feature.

[ci skip-rust]
[ci skip-build-wheels]