Allow using file addresses on the command line (#10573)
Now, you can specify `a/b/c.txt:tgt` on the command line. Pants will interpret this as it would if it were defined in the `dependencies` field of a BUILD file. Pants will treat this as a file address and look up the original base target `a/b/c:tgt`, along with validating that this file actually belongs to the target.

If there is no `:` in the file name, then we will continue to treat this as a filesystem spec and will find the owner(s) of that file.

We expect this to be useful for users feeding the output of other Pants commands into a new run. For example, we now output with `test` that we are running on a file address, so now the user can copy that address and run `./pants test $file_addr`.

This also removes redundant work with `--changed-since` where we were looking up the owner of the same file twice.