Lower the default for `--python-setup-resolver-jobs` (#10584)
See https://github.com/pantsbuild/pants/pull/9253 for the motivation. We didn't land the change then because we had to account for both v1 and v2 usages, and we didn't want to optimize for v2 at the harm of v1. Now we only have v2 to worry about.

This partially closes https://github.com/pantsbuild/pants/issues/9964, although in a much more naive way.

[ci skip-rust]
[ci skip-build-wheels]