Fix binary running to correctly work with source roots. (#10607)
When setting the PEX_EXTRA_SYS_PATH env var, we failed to take the
chroot into account. Now we do.

Unfortunately there was no easy way to plumb the chroot through to
the implementation (as we do for repl), so instead we allow it to be
injected via string formatting. Since "value".format(chroot='path') is
a no-op if the value doesn't contain "{chroot}", this is robust.

For uniformity, we use the same mechanism to add the chroot to the
binary path, and to any other args (e.g., in the future we might want
to reference a file in the args).

Also improves the implementation of python running:

Reuses the standard requirements.pex, so we don't have
to re-resolve just to create the runnable pex. The runner
pex uses PEX_PATH to bring in the requirements pex.
Ignores the binary's platform settings, since run is
always local, so platforms aren't relevant.

[ci skip-rust]

[ci skip-build-wheels]