Add `--python-infer-string-imports` (#10609)
Certain Python projects, such as Django apps, make frequent use of dynamic imports. Dep inference needs to be able to understand those to work well with those code bases.

We turn string imports inference off by default because it could be surprising and it's much more likely to have false positives.

### Rejected design (for now) - global ignores

With Toolchain's buildgen (which inspired dep inference), false positives were handled via a subsystem option that allowed users to put regex strings to ignore globally. At the time, there was no `!` ignore syntax in the `dependencies` field, so this was the only option.

Now that we have `!` ignores, we start with the simplest solution to only use that for false positives. If we get user feedback that they want a decentralized mechanism, we can easily add a subsystem.

(Also note that we will only infer dependencies on Python targets, whereas Toolchain's Buildgen would infer on `files()` and `resources()`. Many of the ignores in Toolchain were for static assets, like `.css` files.)

[ci skip-rust]
[ci skip-build-wheels]