Spawning against materialized binaries works. (#10577)
It's generically unsafe to fork+exec against binaries written out in a
multithreaded program when concurrent forks are possible. Even if all
files are opened for writing with O_CLOEXEC (which is the case for
Rust) if thread1 opens a file for writing and then thread2 forks,
thread2 will hold an open file descriptor. If thread2's subsequent exec
is delayed past the fork+exec thread1 does against the file it wrote,
then the thread1 fork+exec'd process will encounter ETXTBSY.

OSX "solves" this by retrying some number of times when it hits ETXTBSY,
but Linux does not attempt this hack. O_CLOFORK has been proposed and
adopted by some unices, but not Linux.

As such we need to make some tradeoff to allow this use case. This 
change introduces a lock around process spawns (fork+exec) to prevent
interleaved fork / exec whenever there is a spawn that we know exec's a 
binary we wrote out. Since forks can still happen in our process space
outside of our control (in libraries), this change also introduces a 
bounded ETXTBSY retry loop as a fallback for binaries we wrote out.

Fixes #10507