Simplify Process descriptions now that we always run on files
Now that we always run on a generated subtarget, which must have exactly one file, we can output that we are running on files instead of targets. Users don't need to know it's really a target.

This stops enumerating the actual target names because almost always the value gets truncated, which is noisy.

# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]