Allow `test --debug` to run on multiple targets sequentially (#10629)
It's now possible to run `./pants test --debug ::`, which will behave like v1 in that it will run sequentially.

There was no solid reason for us to put a restriction that you had to run one-at-a-time. Yes, this will lose parallelism that we want users to have, but Everyone Is An Adult Here. If the user wants to run sequentially, all power to them.

For now, we continue running if a certain test fails. We do not add a `--fail-early` flag like we had in V1 because that flag would only end up applying to `--test-debug`, rather than the default of `--no-test-debug`, which is confusing. The user can early exit by using `ctrl-c`. (They could use Pytest's `-x`, but this will only apply to each file; it will cause that file to early return, but it won't cause other files to fail due to the failure in another file.)

[ci skip-rust]
[ci skip-build-wheels]