Change level for some more substantial rules to Debug
These rules would actually be helpful when debugging. They a) have lots of unique logic, and b) are not run hundreds of times in MultiGets.

I think it's still the right decision for rules to default to Trace because that is a more sensible default. We default to the rule not being very interesting, like a rule to go from TestResult -> EnrichedTestResult.

This also automatically updates all @goal_rules to be debug level.

# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]