Reduce noisy log messages from `@rule` params (#10658)
### Problem

cf. https://github.com/pantsbuild/pants/issues/10518 , it's possible for pants to emit extremely large log messages whenever it prints a long containing the string representation of large Python types. In the above issue, the large Python types are  `FieldSet` subclasses (which can include lots of e.g. file path names) in their Python `__str__` representation. These types appear as the `params` field on a `Task`, and are part of the Rust `Display` representation of a `Task`, which is part of the log message when a node is invalidated because of a file system change, and might more generally show up in any pants log message.

### Solution

A quick fix for the above problem is to simply remove the `Params` from the `Display` implementation of an `@rule`'s task, and only include the name of the Python function corresponding to the rule. cf. https://github.com/pantsbuild/pants/issues/7907 ,  we probably want to eventually implement a more sophisticated way of deciding what Python type information we do and do not need to include in useful pants logs from Rust.