Add transitive excludes for `dependencies` with `!!` (#10663)
## Problem

Closes https://github.com/pantsbuild/pants/issues/10661. As described there, sometimes users want to be able to exclude certain problematic dependencies from things like `python_binary` or `python_tests`, such as TensorFlow. Direct dependency ignores via `!` are not adequate because those dependencies might be a transitive dependency.

The user could instead be disciplined with their transitive dependencies to make sure the dep they don't want is not used. However, this is difficult to get right. It would also often mean creating two targets for the same library code, one with the dep and one without; that would mean dependency inference couldn't work anymore due to ambiguous owners.

## Solution

For target types that support it, you can add `!!path/to:address` to the `dependencies` field. This will ignore the dependency both if it's a direct dependency and if it's a transitive dependency.

To limit the risk of confusion, this new mechanism only works with certain target types that have declared they support the syntax, such as `python_binary`. These target types are meant to be "roots", i.e. not have dependees. Otherwise, users get this error:

```
▶ ./pants dependencies src/python/pants/util
15:59:08 [ERROR] 1 Exception encountered:

  TransitiveExcludesNotSupportedError: Bad value '!!//:pants_toml' in the `dependencies` field for src/python/pants/util. Transitive excludes with `!!` are not supported for this target type. Did you mean to use a single `!` for a direct exclude?

Transitive excludes work with these target types: ['python_binary', 'python_distribution', 'python_tests']

(Use --print-exception-stacktrace to see more error details.)
```

[ci skip-rust]
[ci skip-build-wheels]