Add an FFI function for reading a `FrozenDict` (#10666)
### Problem

We want to be able to convert the relevant parts of a `Process` to an `InteractiveProcess`. This will help, for example, so that we can use the normal Pytest `Process` to populate the `env` and `argv` for an `InteractiveProcess`. However, this is really tricky to do because we store `Process.env` as a `Tuple[str, ...]`, rather than a `Mapping[str, str]` or `Tuple[Tuple[str, str], ...]`. Instead, we ideally want to store the `env` as a `FrozenDict[str, str]`.

### Solution

Teach Rust how a `FrozenDict` works, which is really a light wrapper around a dictionary. This means we can use the CPython crate's `PyDict` type, then convert that to a `BTreeMap`.

[ci skip-build-wheels]