Simplify `python_requirement_library` by deprecating the `python_requirement` object (#10669)
### Problem

An inline Python requirement object involves lots of boilerplates, in large part due to the `python_requirement` object. Currently, we only ever use the requirement string and the `modules` field belonging to that object; all other fields are ignored.

Beyond adding to boilerplate, having to deal with `python_requirement` makes it harder for us to add robust support for Poetry and Pipenv because it makes our abstractions more complex than necessary.

### Solution

Deprecate `python_requirement` and instead use a sequence of strings in the `requirements` field for a `python_requirement_library`. Add a new field `module_mapping` to it that allows for overriding module names for the sake of dependency inference.

We still allow `python_requirement` objects until 2.1.0 for a smoother migration.

This also refactors and improves the `python_requirements()` macro, including adding nice error messages for invalid values. We also now have comprehensive tests for this macro.

[ci skip-rust]
[ci skip-build-wheels]