Provide better debugging information for rule parameters (#10665)
### Problem

cf. https://github.com/pantsbuild/pants/issues/10334 , the `Engine traceback` strings don't provide enough information about the arguments to the `@rule`s they represent to be useful for debugging.

### Solution

This commit introduces a new "EngineAware" method `debug_hint`, which can be implemented for the parameters of an `@rule` (cf. https://github.com/pantsbuild/pants/issues/7907#issuecomment-648463655 ) to provide a concise debugging string. This is implemented for the `TestFieldSet` type, writing the address of the test in question, and the value of this method is plumbed to the `EngineTraceback`  string. This commit also splits up the `EngineAware` class into two separate classes: `EngineAwareParams` to be implemented by input parameters to a rule; and `EngineAwareReturnType` to be implemented by rule return types.

Additionally this commit does a bit of cleanup to the string representation of the `Params` type, and puts the code for manipulating `EngineAware` methods into a new submodule.

### Result

https://asciinema.org/a/354849 demonstrates the expanded information in `Engine traceback` strings.