Handle workunits corresponding to canceled Nodes. (#10659)
### Problem

cf. https://github.com/pantsbuild/pants/issues/10650 . When a node is canceled by the engine, the engine will remove the future associated with that node without communicating this fact to the workunit store. This means that the store has no idea that a Started but not yet Completed workunit is never going to complete.

### Solution

This commit adds a `CanceledWorkunitGuard` in the `with_workunit` public interface to workunits. If a workunit completes normally, the last thing it will do is call `not_canceled()` on the guard, and the guard will do nothing. If the guard is dropped before this happens, though, its `Drop` implementation will call a cleanup function on the workunit store to remove the workunit from the heavy hitters and display a log message stating that the workunit has been canceled.