Fix `run`, `repl`, and `test --debug` to have hermetic environments (#10668)
Previously, `InteractiveProcess` used whatever the `env` was for the `./pants` process. While this can be convenient, it is not consistent with the rest of Pants. For example, it was possible for a test to pass with `test --debug` but fail with `test --no-debug`, or vice versa, due to differences with the env.

This change reinforces the need for https://github.com/pantsbuild/pants/issues/10644, and for that to be ergonomic for users to set arbitrary env vars.

[ci skip-build-wheels]