Change `pants_integration_test.py` to use Pytest-style functions (#10675)
We've slowly been chipping away at using Pytest-style tests, meaning using `assert` rather than `self.assertX` and using top-level functions rather than test classes.

Now, integration tests can be written in Pytest style. We deprecate the old class-based style.

To land this change, we drop the ability to set the PANTS_PROFILE in an integration test. It appears that this was not used; the workaround is to run the test like you would on the command line.

[ci skip-build-wheels]